+++ AdminPanel/AdminPanel/CustomScripts/jsAssignPageToUser.js
@@ -75,7 +75,8 @@ function assign_page_list() {
             //alert(responseFromServer);
             var jsonData = eval('(' + responseFromServer + ')');
             //alert(jsonData[0].index_id);
-            var tbl = "<table>";
+            var tbl = "<table width='100%'>";
+            tbl += "<tr><th>Id</th><th>Page Name</th><th>Delete</th><th>Update</th></tr>";
             for (var iCount = 0; iCount < jsonData.length; iCount++) {
                 tbl += "<tr>";
                 tbl += "<td>";
@@ -85,10 +86,10 @@ function assign_page_list() {
                 tbl += jsonData[iCount].page_id;
                 tbl += "</td>";
                 tbl += "<td>";
-                tbl += "<input type='button' value='DELETE' onclick='del(" + jsonData[iCount].index_id + ")'>";
+                tbl += "<i type='button' class='fa fa-trash btn-danger'  onclick='del(" + jsonData[iCount].index_id + ")'></i>";
                 tbl += "</td>";
                 tbl += "<td>";
-                tbl += "<input type='button' value='UPDATE' onclick='update(" + jsonData[iCount].index_id + ")'>";
+                tbl += "<i type='button' class='fa fa-edit btn-warning' onclick='update(" + jsonData[iCount].index_id + ")'></i>";
                 tbl += "</td>";
                 tbl += "</tr>";
             }
@@ -132,20 +133,22 @@ function update(id) {
 
 
 function del(id) {
-    $.ajax({
-        //url: "https://trial.spyderxindia.com/api/assignpagestouser/3",
-        type: "DELETE",
-        //crossOrigin: true,
-        url: "https://trial.spyderxindia.com/api/assignpagestouser/" + id,
-        //data: JSON.stringify(num),
-        contentType: "application/json",
-        //datatype: "json",
-        success: function (responseFromServer) {
-            alert(responseFromServer);
-            assign_page_list();
+    if (confirm('Are you sure ?')) {
+        $.ajax({
+            //url: "https://trial.spyderxindia.com/api/assignpagestouser/3",
+            type: "DELETE",
+            //crossOrigin: true,
+            url: "https://trial.spyderxindia.com/api/assignpagestouser/" + id,
+            //data: JSON.stringify(num),
+            contentType: "application/json",
+            //datatype: "json",
+            success: function (responseFromServer) {
+                alert(responseFromServer);
+                assign_page_list();
 
 
-        }
-    });
+            }
+        });
+    }
 
 }
