+++ AdminPanel/AdminPanel/CustomScripts/jsPlanTypeMst.js
@@ -3,18 +3,28 @@
     plan_type_list();
 
     $("#btnSave").click(function () {
-        var requestData = '[{"plan_type_name":"' + $("#txtPlan_Type_Name").val() + '","user_id":"11"}]';
-        $.ajax({
-            type: "POST",
-            url: "https://trial.spyderxindia.com/api/plantypemst",
-            data: JSON.stringify(requestData),
-            contentType: "application/json",
-            //datatype: "json",
-            success: function (responseFromServer) {
-                alert(responseFromServer);
-                plan_type_list();
-            }
-        });
+
+        var re = /^[a-zA-Z]+$/;
+        if (!re.test($("#txtPlan_Type_Name").val())) {
+
+            alert("Enter valid Plan Type Name...!");
+
+        }
+        else {
+
+            var requestData = '[{"plan_type_name":"' + $("#txtPlan_Type_Name").val() + '","user_id":"11"}]';
+            $.ajax({
+                type: "POST",
+                url: "https://trial.spyderxindia.com/api/plantypemst",
+                data: JSON.stringify(requestData),
+                contentType: "application/json",
+                //datatype: "json",
+                success: function (responseFromServer) {
+                    alert(responseFromServer);
+                    plan_type_list();
+                }
+            });
+        }
     });
 
 
@@ -133,20 +143,21 @@ function update(id) {
 
 
 function del(id) {
-    $.ajax({
-        //url: "https://trial.spyderxindia.com/api/plantypemst/3",
-        type: "DELETE",
-        //crossOrigin: true,
-        url: "https://trial.spyderxindia.com/api/plantypemst/" + id,
-        //data: JSON.stringify(num),
-        contentType: "application/json",
-        //datatype: "json",
-        success: function (responseFromServer) {
-            alert(responseFromServer);
-            plan_type_list();
+    if (confirm('Are you sure ?')) {
+        $.ajax({
+            //url: "https://trial.spyderxindia.com/api/plantypemst/3",
+            type: "DELETE",
+            //crossOrigin: true,
+            url: "https://trial.spyderxindia.com/api/plantypemst/" + id,
+            //data: JSON.stringify(num),
+            contentType: "application/json",
+            //datatype: "json",
+            success: function (responseFromServer) {
+                alert(responseFromServer);
+                plan_type_list();
 
 
-        }
-    });
-
+            }
+        });
+    }
 }
